--- src/main/java/com/hantsylabs/restexample/springmvc/config/Jackson2ObjectMapperConfig.java
@@ -1,34 +1,0 @@
-
-package com.hantsylabs.restexample.springmvc.config;
-
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
-
-/**
- *
- * @author Hantsy Bai<hantsy@gmail.com>
- */
-@Configuration
-public class Jackson2ObjectMapperConfig {
-
-    @Bean
-    @Primary
-    public ObjectMapper objectMapper() {
-
-        Jackson2ObjectMapperBuilder builder = new Jackson2ObjectMapperBuilder();
-        builder.serializationInclusion(Include.NON_EMPTY);
-        builder.featuresToDisable(
-               // SerializationFeature.WRITE_DATES_AS_TIMESTAMPS,
-                DeserializationFeature.FAIL_ON_IGNORED_PROPERTIES,
-                DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
-        builder.featuresToEnable(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY);
-
-        return builder.build();
-    }
-
-}
