--- src/main/java/com/hantsylabs/restexample/springmvc/config/MessageSourceConfig.java
@@ -1,19 +1,0 @@
-package com.hantsylabs.restexample.springmvc.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-
-@Configuration
-public class MessageSourceConfig {
-
-    @Bean
-    public ReloadableResourceBundleMessageSource messageSource() {
-        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-        messageSource.setDefaultEncoding("UTF-8");
-        messageSource.setBasenames("classpath:i18n/messages", "classpath:i18n/errors");
-
-        return messageSource;
-    }
-
-}
