--- src/main/java/com/hantsylabs/restexample/springmvc/config/WebConfig.java
@@ -1,76 +1,0 @@
-package com.hantsylabs.restexample.springmvc.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import javax.inject.Inject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.MediaType;
-import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-import org.springframework.web.servlet.mvc.method.annotation.ExceptionHandlerExceptionResolver;
-
-@Configuration
-@EnableWebMvc
-public class WebConfig extends WebMvcConfigurerAdapter {
-
-    private static final Logger logger = LoggerFactory.getLogger(WebConfig.class);
-
-    @Inject
-    private ObjectMapper objectMapper;
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        
-        registry.addResourceHandler("/**")
-            .addResourceLocations("classpath:/META-INF/resources/");
-
-    }
-
-    @Override
-    public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addRedirectViewController("/", "/index.html");
-    }
-    
-    
-
-    @Override
-    public void configureHandlerExceptionResolvers(List<HandlerExceptionResolver> exceptionResolvers) {
-        exceptionResolvers.add(exceptionHandlerExceptionResolver());
-    }
-
-    @Override
-    public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
-        List<HttpMessageConverter<?>> messageConverters = messageConverters();
-        converters.addAll(messageConverters);
-    }
-
-    @Bean
-    public ExceptionHandlerExceptionResolver exceptionHandlerExceptionResolver() {
-        ExceptionHandlerExceptionResolver exceptionHandlerExceptionResolver = new ExceptionHandlerExceptionResolver();
-        exceptionHandlerExceptionResolver.setMessageConverters(messageConverters());
-
-        return exceptionHandlerExceptionResolver;
-    }
-
-    private List<HttpMessageConverter<?>> messageConverters() {
-        List<HttpMessageConverter<?>> messageConverters = new ArrayList<>();
-
-        MappingJackson2HttpMessageConverter jackson2Converter = new MappingJackson2HttpMessageConverter();
-        jackson2Converter.setSupportedMediaTypes(Arrays.asList(MediaType.APPLICATION_JSON));
-        jackson2Converter.setObjectMapper(objectMapper);
-
-        messageConverters.add(jackson2Converter);
-        return messageConverters;
-    }
-
-}
