+++ src/main/resources/META-INF/resources/i18n/messages_en.properties
@@ -1,5 +1,6 @@
 appName=REST EXAMPLE APP
 search=Search
+add=Add
 personalHome=Personal Home Page
 username=Username
 password=Password
