+++ src/main/resources/META-INF/resources/i18n/messages_zh_CN.properties
@@ -1,10 +1,11 @@
 appName=REST \u5b9e\u4f8b
 search=\u67e5\u8be2
+add=\u6dfb\u52a0
 
 personalHome=\u4e2a\u4eba\u8d44\u6599
 username=\u7528\u6237\u540d
 password=\u5bc6\u7801
-name=\u59d3\u540d
+displayName=\u59d3\u540d
 role=\u89d2\u8272
 save=\u4fdd\u5b58
 cancel=\u53d6\u6d88
