+++ src/main/resources/META-INF/resources/partials/posts/home.html
@@ -53,7 +53,7 @@
                 <tr ng-repeat="post in posts|orderBy:-createdOn">
                     <td><a ng-href="#/posts/{{post.id}}">{{post.title}}</a></td>
                     <td>{{post.status}}</td>
-                    <td>{{post.createdDate|date:'short'}}</td>
+                    <td>{{post.createdDate}}</td>
                     <td>{{post.createdBy.name}}</td>
                 </tr>
             </table>
